requirements.txt
# ===============================
# Core LLM / Agent Stack
# ===============================
langchain==0.3.25
langchain-core==0.3.83
langchain-openai==0.3.16
langchain-community==0.3.24
langchain-text-splitters==0.3.8
langgraph==0.3.34

openai==1.109.1
tiktoken==0.9.0

# ===============================
# RAG / Vector
# ===============================
faiss-cpu==1.11.0
pypdf==5.4.0
docx2txt==0.8

# ===============================
# Sentiment
# ===============================
vaderSentiment==3.3.2

# ===============================
# UI
# ===============================
streamlit==1.54.0

# ===============================
# API Layer
# ===============================
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20

# ===============================
# Utilities
# ===============================
python-dotenv==1.1.0
pydantic==2.11.4
pydantic-settings==2.9.1
aiofiles==24.1.0
httpx==0.28.1
requests==2.32.3

# ===============================
# SEC + Scraping
# ===============================
sec-edgar-downloader==5.1.0
beautifulsoup4==4.12.3
lxml==5.3.0
newspaper3k==0.2.8

# ===============================
# Critical Compatibility Pin
# ===============================
packaging<26.0.0